+++ src/inputs/getters.ts
@@ -5,6 +5,7 @@ import type { InputValue } from "./types";
 import {
   INPUT_NAME_IGNORE,
   INPUT_NAME_DRY_RUN,
+  INPUT_NAME_STRICT,
   INPUT_NAME_SVGO_CONFIG,
   INPUT_NAME_SVGO_VERSION,
 } from "../constants";
@@ -118,7 +119,26 @@ function getIsDryRun(
   const inputName = INPUT_NAME_DRY_RUN;
   const input = getBooleanInput({ inp, inputName, defaultValue });
   if (!input.valid) {
-    return [input, errors.New("invalid dry-run value")];
+    return [
+      { ...input, value: true },
+      errors.New("invalid dry-run value"),
+    ];
+  }
+
+  return [input, null];
+}
+
+function getIsStrictMode(
+  inp: Inputter,
+  defaultValue: boolean,
+): [InputValue<boolean>, error] {
+  const inputName = INPUT_NAME_STRICT;
+  const input = getBooleanInput({ inp, inputName, defaultValue });
+  if (!input.valid) {
+    return [
+      { ...input, value: true },
+      errors.New("invalid strict value"),
+    ];
   }
 
   return [input, null];
@@ -157,6 +177,7 @@ function getSvgoVersion(
 export {
   getIgnoreGlobs,
   getIsDryRun,
+  getIsStrictMode,
   getSvgoConfigPath,
   getSvgoVersion,
 };